requirements.txt
streamlit
PyMuPDF
python-dotenv
requests
